+++ plugin.video.wral/addon.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
 <addon id="plugin.video.wral"
        name="WCBS Programs"
-       version="3.0.1"
+       version="3.0.3"
        provider-name="t1m">
   <requires>
     <import addon="xbmc.python" version="2.20.0"/>
-    <import addon="script.module.t1mlib" version="1.0.14"/>
+    <import addon="script.module.t1mlib" version="1.0.15"/>
   </requires>
   <extension point="xbmc.python.pluginsource"
             library="default.py">
@@ -21,6 +21,6 @@
     <forum>http://forum.kodi.tv/showthread.php?tid=222072</forum>
     <website>www.wral.com</website>
     <email></email>
-    <source>https://github.com/learningit/plugin.video.wral/</source>
+    <source>https://github.com/learningit/Kodi-plugins-source</source>
   </extension>
 </addon>
