+++ plugin.video.wral/changelog.txt
@@ -1,2 +1,4 @@
-V2.0.1 Initial version
+V3.0.3 more cleanup
+V3.0.2 fix for more than 100 shows; add evening news
 V3.0.1 Separate scraper for future functions
+V2.0.1 Initial version
