+++ plugin.video.wral/readme.txt
@@ -1,7 +1,16 @@
 plugin.video.wral
-================
+=====================
+Kodi Addon for WCBS Programs website
 
-Kodi Addon for WRAL (WCBS Programs) website
+icon.png sourced from public domain : 
+https://commons.wikimedia.org/wiki/File:CBS_classic_logo.svg
 
+fanart.jpg sourced from public domain:
+https://commons.wikimedia.org/wiki/File:CBS_classic_logo.svg (converted to .jpg)
+
+Change Log
+=====================
+V3.0.3 more cleanup
+V3.0.2 fix for more than 100 shows; add evening news
 V3.0.1 Separate scraper for future functions
 V2.0.1 Initial version
